Move relocation artifact into Veryl frontend
