+++ crates/celox/src/flatting.rs
@@ -87,8 +87,6 @@ pub fn flatting(
     }
 
     Ok(RelocationModule {
-        #[cfg(test)]
-        variables: module.variables.clone(),
         eval_apply_ff_blocks: HashMap::default(),
         eval_only_ff_blocks: HashMap::default(),
         apply_ff_blocks: HashMap::default(),
